@@ -1,3 +1,6 @@
+/**
+ * Imports for file system operations, diff viewing, and utility functions
+ */
 import * as path from "path"
 import { DiffViewProvider } from "../../../../../integrations/editor/diff-view-provider"
 import { ClaudeSayTool } from "../../../../../shared/ExtensionMessage"
@@ -6,21 +9,66 @@ import { BaseAgentTool } from "../../base-agent.tool"
 import { AgentToolOptions, AgentToolParams } from "../../types"
 import fs from "fs"
 import { detectCodeOmission } from "./detect-code-omission"
-import { parseDiffBlocks, applyEditBlocksToFile, checkFileExists, preprocessContent } from "./utils"
+import { parseDiffBlocks, applyEditBlocksToFile, checkFileExists, preprocessContent, EditBlock } from "./utils"
+import { InlineEditHandler } from "../../../../../integrations/editor/inline-editor"
+import { ToolResponseV2 } from "../../../../../agent/v1/types"
+import delay from "delay"
+import PQueue from "p-queue"
 
+/**
+ * FileEditorTool handles file editing operations in the VSCode extension.
+ * It provides functionality for:
+ * - Showing diffs between file versions
+ * - Handling partial updates during file editing
+ * - Managing inline edits with search/replace blocks
+ * - Processing user approvals for file changes
+ *
+ * The tool supports both complete file rewrites and partial edits through diff blocks.
+ * It includes features for animation control and update throttling to ensure smooth performance.
+ */
 export class FileEditorTool extends BaseAgentTool {
+	/** Tool parameters including update callbacks and input */
 	protected params: AgentToolParams
+	/** Provider for showing file diffs in VSCode */
 	public diffViewProvider: DiffViewProvider
+	/** Handler for inline file editing operations */
+	public inlineEditor: InlineEditHandler
+	/** Flag indicating if final content is being processed */
 	private isProcessingFinalContent: boolean = false
+	/** Timestamp of last update for throttling */
 	private lastUpdateTime: number = 0
+	/** Minimum interval between updates in milliseconds */
 	private readonly UPDATE_INTERVAL = 16
+	/** Queue for processing partial updates sequentially */
+	private pQueue: PQueue = new PQueue({ concurrency: 1 })
+	/** Flag to control write animation display */
 	private skipWriteAnimation: boolean = false
+	/** Counter for tracking update sequence */
 	private updateNumber: number = 0
+	/** Array of edit blocks being processed */
+	private editBlocks: EditBlock[] = []
+	/** Current file state including path and content */
+	private fileState?: {
+		absolutePath: string
+		orignalContent: string
+		isExistingFile: boolean
+	}
+	/** ID of the last applied edit block */
+	private lastAppliedEditBlockId: string = ""
 
+	/**
+	 * Initializes the FileEditorTool with required parameters and options.
+	 * Sets up diff view provider and inline editor handler.
+	 * Configures animation settings based on state manager.
+	 *
+	 * @param params - Tool parameters including callbacks and input
+	 * @param options - Configuration options for the tool
+	 */
 	constructor(params: AgentToolParams, options: AgentToolOptions) {
 		super(options)
 		this.params = params
 		this.diffViewProvider = new DiffViewProvider(getCwd(), this.koduDev)
+		this.inlineEditor = new InlineEditHandler()
 		if (!!this.koduDev.getStateManager().skipWriteAnimation) {
 			this.skipWriteAnimation = true
 		}
@@ -31,15 +79,13 @@ export class FileEditorTool extends BaseAgentTool {
 		return result
 	}
 
-	public async handlePartialUpdateDiff(relPath: string, diff: string): Promise<void> {
+	public async _handlePartialUpdateDiff(relPath: string, diff: string): Promise<void> {
 		// this might happen because the diff view are not instant.
 		if (this.isProcessingFinalContent) {
 			this.logger("Skipping partial update because the tool is processing the final content.", "warn")
 			return
 		}
-		// if the user has skipped the write animation, we don't need to show the diff view until we reach the final state
-		// if (this.skipWriteAnimation) {
-		await this.params.updateAsk(
+		this.params.updateAsk(
 			"tool",
 			{
 				tool: {
@@ -53,44 +99,122 @@ export class FileEditorTool extends BaseAgentTool {
 			},
 			this.ts
 		)
-		return
-		// }
 
-		const currentTime = Date.now()
-		// don't push too many updates to the diff view provider to avoid performance issues
-		if (currentTime - this.lastUpdateTime < this.UPDATE_INTERVAL) {
+		if (!diff.includes("REPLACE")) {
+			this.logger("Skipping partial update because the diff does not contain REPLACE keyword.", "warn")
 			return
 		}
-
-		if (!this.diffViewProvider.isDiffViewOpen()) {
+		// if the user has skipped the write animation, we don't need to show the diff view until we reach the final state
+		if (this.skipWriteAnimation) {
+			return
+		}
+		try {
+			this.editBlocks = parseDiffBlocks(diff, this.fileState!.absolutePath)
+		} catch (err) {
+			this.logger(`Error parsing diff blocks: ${err}`, "error")
+			return
+		}
+		if (!this.inlineEditor.isOpen()) {
 			try {
-				// this actually opens the diff view but might take an extra few ms to be considered open requires interval check
-				// it can take up to 300ms to open the diff view
-				await this.diffViewProvider.open(relPath)
+				await this.inlineEditor.open(
+					this.editBlocks[0].id,
+					this.fileState!.absolutePath,
+					this.editBlocks[0].searchContent
+				)
 			} catch (e) {
 				this.logger("Error opening diff view: " + e, "error")
 				return
 			}
 		}
-		const absolutePath = path.resolve(getCwd(), relPath)
-		const fileExists = await checkFileExists(relPath)
-		if (!fileExists) {
-			throw new Error("File does not exist, but 'diff' parameter is provided")
+		// now we are going to start applying the diff blocks
+		if (this.editBlocks.length > 0) {
+			const currentBlock = this.editBlocks.at(-1)
+			if (!currentBlock?.replaceContent) {
+				return
+			}
+
+			// If this block hasn't been tracked yet, initialize it
+			if (!this.editBlocks.some((block) => block.id === currentBlock.id)) {
+				// Clean up any SEARCH text from the last block before starting new one
+				if (this.lastAppliedEditBlockId) {
+					const lastBlock = this.editBlocks.find((block) => block.id === this.lastAppliedEditBlockId)
+					if (lastBlock) {
+						const lines = lastBlock.replaceContent.split("\n")
+						// Only remove the last line if it ONLY contains a partial SEARCH
+						if (lines.length > 0 && /^=?=?=?=?=?=?=?$/.test(lines[lines.length - 1].trim())) {
+							lines.pop()
+							await this.inlineEditor.applyFinalContent(
+								lastBlock.id,
+								lastBlock.searchContent,
+								lines.join("\n")
+							)
+						} else {
+							await this.inlineEditor.applyFinalContent(
+								lastBlock.id,
+								lastBlock.searchContent,
+								lastBlock.replaceContent
+							)
+						}
+					}
+				}
+
+				await this.inlineEditor.open(currentBlock.id, this.fileState!.absolutePath, currentBlock.searchContent)
+				this.editBlocks.push({
+					id: currentBlock.id,
+					replaceContent: currentBlock.replaceContent,
+					path: this.fileState!.absolutePath,
+					searchContent: currentBlock.searchContent,
+				})
+				this.lastAppliedEditBlockId = currentBlock.id
+			}
+
+			const blockData = this.editBlocks.find((block) => block.id === currentBlock.id)
+			if (blockData) {
+				blockData.replaceContent = currentBlock.replaceContent
+				await this.inlineEditor.applyStreamContent(
+					currentBlock.id,
+					currentBlock.searchContent,
+					currentBlock.replaceContent
+				)
+			}
 		}
 
-		// Read existing file content
-		const originalContent = await fs.promises.readFile(absolutePath, "utf-8")
-
-		try {
-			// Parse and apply the edit blocks
-			const editBlocks = parseDiffBlocks(diff, absolutePath)
-			this.logger(`Parsed edit blocks: ${JSON.stringify(editBlocks)}`, "debug")
-			const newContent = await applyEditBlocksToFile(originalContent, editBlocks)
-			await this.diffViewProvider.update(newContent, false)
-			this.lastUpdateTime = currentTime
-		} catch (e) {
-			this.logger(`Not enough information to update the diff view: ${e}`, "warn")
+		// Finalize the last block
+		if (this.lastAppliedEditBlockId) {
+			const lastBlock = this.editBlocks.find((block) => block.id === this.lastAppliedEditBlockId)
+			if (lastBlock) {
+				const lines = lastBlock.replaceContent.split("\n")
+				await this.inlineEditor.applyFinalContent(lastBlock.id, lastBlock.searchContent, lines.join("\n"))
+			}
 		}
+	}
+
+	public async handlePartialUpdateDiff(relPath: string, diff: string): Promise<void> {
+		if (!this.fileState) {
+			const absolutePath = path.resolve(getCwd(), relPath)
+			const isExistingFile = await checkFileExists(relPath)
+			if (!isExistingFile) {
+				this.logger(`File does not exist: ${relPath}`, "error")
+				this.fileState = {
+					absolutePath,
+					orignalContent: "",
+					isExistingFile: false,
+				}
+				return
+			}
+			const originalContent = fs.readFileSync(absolutePath, "utf8")
+			this.fileState = {
+				absolutePath,
+				orignalContent: originalContent,
+				isExistingFile,
+			}
+		}
+		// this might happen because the diff view are not instant.
+		if (this.isProcessingFinalContent) {
+			this.logger("Skipping partial update because the tool is processing the final content.", "warn")
+			return
+		}
+		this.pQueue.add(() => this._handlePartialUpdateDiff(relPath, diff))
 	}
 
 	/**
@@ -100,29 +224,46 @@ export class FileEditorTool extends BaseAgentTool {
 	 * @returns
 	 */
 	public async handlePartialUpdate(relPath: string, acculmatedContent: string): Promise<void> {
+		if (!this.fileState) {
+			const absolutePath = path.resolve(getCwd(), relPath)
+			const isExistingFile = await checkFileExists(relPath)
+			if (!isExistingFile) {
+				this.logger(`File does not exist: ${relPath}`, "error")
+				this.fileState = {
+					absolutePath,
+					orignalContent: "",
+					isExistingFile: false,
+				}
+			} else {
+				const originalContent = fs.readFileSync(absolutePath, "utf8")
+				this.fileState = {
+					absolutePath,
+					orignalContent: originalContent,
+					isExistingFile,
+				}
+			}
+		}
 		// this might happen because the diff view are not instant.
 		if (this.isProcessingFinalContent) {
 			this.logger("Skipping partial update because the tool is processing the final content.", "warn")
 			return
 		}
+
 		this.updateNumber++
 		// if the user has skipped the write animation, we don't need to show the diff view until we reach the final state
-		if (this.skipWriteAnimation) {
-			await this.params.updateAsk(
-				"tool",
-				{
-					tool: {
-						tool: "write_to_file",
-						content: acculmatedContent,
-						path: relPath,
-						ts: this.ts,
-						approvalState: "loading",
-					},
+		await this.params.updateAsk(
+			"tool",
+			{
+				tool: {
+					tool: "write_to_file",
+					content: acculmatedContent,
+					path: relPath,
+					ts: this.ts,
+					approvalState: "loading",
 				},
-				this.ts
-			)
-			return
-		}
+			},
+			this.ts
+		)
 
 		const currentTime = Date.now()
 		// don't push too many updates to the diff view provider to avoid performance issues
@@ -144,6 +285,200 @@ export class FileEditorTool extends BaseAgentTool {
 		this.lastUpdateTime = currentTime
 	}
 
+	private async finalizeInlineEdit(path: string, content: string): Promise<ToolResponseV2> {
+		// we are going to parse the content and apply the changes to the file
+		this.isProcessingFinalContent = true
+		try {
+			this.editBlocks = parseDiffBlocks(content, path)
+		} catch (err) {
+			this.logger(`Error parsing diff blocks: ${err}`, "error")
+			throw new Error(`Error parsing diff blocks: ${err}`)
+		}
+		if (!this.inlineEditor.isOpen()) {
+			this.inlineEditor.open(
+				this.editBlocks[0].id,
+				this.fileState?.absolutePath!,
+				this.editBlocks[0].searchContent
+			)
+		}
+		await this.inlineEditor.forceFinalizeAll(this.editBlocks!)
+		// now we are going to prompt the user to approve the changes
+		const { response, text, images } = await this.params.ask(
+			"tool",
+			{
+				tool: {
+					tool: "write_to_file",
+					content,
+					approvalState: "pending",
+					path,
+					ts: this.ts,
+				},
+			},
+			this.ts
+		)
+		if (response !== "yesButtonTapped") {
+			await this.params.updateAsk(
+				"tool",
+				{
+					tool: {
+						tool: "write_to_file",
+						content,
+						approvalState: "rejected",
+						path,
+						ts: this.ts,
+						userFeedback: text,
+					},
+				},
+				this.ts
+			)
+			// revert the changes
+			await this.inlineEditor.rejectChanges()
+			if (response === "noButtonTapped") {
+				return this.toolResponse("rejected", "Write operation cancelled by user.")
+			}
+			// If not a yes or no, the user provided feedback (wrote in the input)
+			await this.params.say("user_feedback", text ?? "The user denied this operation.", images)
+			return this.toolResponse(
+				"feedback",
+				`The user denied this operation with the following feedback:\n<user_feedback>${
+					text ?? "No text feedback provided. check the images for more details."
+				}</user_feedback>`,
+				images
+			)
+		}
+		await this.inlineEditor.saveChanges()
+		this.koduDev.getStateManager().addErrorPath(path)
+
+		// Final approval state
+		await this.params.updateAsk(
+			"tool",
+			{
+				tool: {
+					tool: "write_to_file",
+					content,
+					approvalState: "approved",
+					path,
+					ts: this.ts,
+				},
+			},
+			this.ts
+		)
+
+		return this.toolResponse("success", "The user approved the changes.")
+	}
+
+	private async finalizeFileEdit(relPath: string, content: string): Promise<ToolResponseV2> {
+		// Show changes in diff view
+		await this.showChangesInDiffView(relPath, content)
+		this.logger(`Changes shown in diff view`, "debug")
+
+		this.logger(`Asking for approval to write to file: ${relPath}`, "info")
+		const { response, text, images } = await this.params.ask(
+			"tool",
+			{
+				tool: {
+					tool: "write_to_file",
+					content: content,
+					approvalState: "pending",
+					path: relPath,
+					ts: this.ts,
+				},
+			},
+			this.ts
+		)
+
+		if (response !== "yesButtonTapped") {
+			await this.params.updateAsk(
+				"tool",
+				{
+					tool: {
+						tool: "write_to_file",
+						content: content,
+						approvalState: "rejected",
+						path: relPath,
+						ts: this.ts,
+						userFeedback: text,
+					},
+				},
+				this.ts
+			)
+			await this.diffViewProvider.revertChanges()
+
+			if (response === "noButtonTapped") {
+				// return formatToolResponse("Write operation cancelled by user.")
+				// return this.toolResponse("rejected", "Write operation cancelled by user.")
+				return this.toolResponse("rejected", "Write operation cancelled by user.")
+			}
+			// If not a yes or no, the user provided feedback (wrote in the input)
+			await this.params.say("user_feedback", text ?? "The user denied this operation.", images)
+			// return formatToolResponse(
+			// 	`The user denied the write operation and provided the following feedback: ${text}`
+			// )
+			return this.toolResponse("feedback", text ?? "The user denied this operation.", images)
+		}
+
+		this.logger(`User approved to write to file: ${relPath}`, "info")
+
+		this.logger(`Saving changes to file: ${relPath}`, "info")
+		// Save changes and handle user edits
+		const { userEdits, finalContent } = await this.diffViewProvider.saveChanges()
+		this.logger(`Changes saved to file: ${relPath}`, "info")
+		this.koduDev.getStateManager().addErrorPath(relPath)
+
+		// Final approval state
+		await this.params.updateAsk(
+			"tool",
+			{
+				tool: {
+					tool: "write_to_file",
+					content: content,
+					approvalState: "approved",
+					path: relPath,
+					ts: this.ts,
+				},
+			},
+			this.ts
+		)
+		this.logger(`Final approval state set for file: ${relPath}`, "info")
+		if (userEdits) {
+			await this.params.say(
+				"user_feedback_diff",
+				JSON.stringify({
+					tool: this.fileState?.isExistingFile ? "editedExistingFile" : "newFileCreated",
+					path: getReadablePath(getCwd(), relPath),
+					diff: userEdits,
+				} as ClaudeSayTool)
+			)
+			return this.toolResponse(
+				"success",
+				`The user made the following updates to your content:\n\nThe updated content has been successfully saved to ${relPath.toPosix()}
+					Here is the latest file content:
+					\`\`\`
+					${finalContent}
+					\`\`\`
+					`
+			)
+		}
+
+		let toolMsg = `The content was successfully saved to ${relPath.toPosix()}.
+			Here is the latest file content:
+			\`\`\`
+			${finalContent}
+			\`\`\`
+			`
+		if (detectCodeOmission(this.diffViewProvider.originalContent || "", finalContent)) {
+			this.logger(`Truncated content detected in ${relPath} at ${this.ts}`, "warn")
+			toolMsg = `The content was successfully saved to ${relPath.toPosix()},
+				but it appears that some code may have been omitted. In caee you didn't write the entire content and included some placeholders or omitted critical parts, please try again with the full output of the code without any omissions / truncations anything similar to "remain", "remains", "unchanged", "rest", "previous", "existing", "..." should be avoided.
+				Here is the latest file content:
+				\`\`\`
+				${finalContent}
+				\`\`\``
+		}
+
+		return this.toolResponse("success", toolMsg)
+	}
+
 	private async processFileWrite() {
 		try {
 			const { path: relPath, kodu_content: content, kodu_diff: diff } = this.params.input
@@ -154,138 +489,15 @@ export class FileEditorTool extends BaseAgentTool {
 
 			// Switch to final state ASAP
 			this.isProcessingFinalContent = true
+			// wait for the queue to be idle
+			await this.pQueue.onIdle()
 
-			const absolutePath = path.resolve(getCwd(), relPath)
-			const fileExists = await checkFileExists(relPath)
-
-			let newContent: string
-
-			if (fileExists && diff) {
-				this.logger(`File exists and diff is provided.`, "debug")
-				// Read existing file content
-				const originalContent = await fs.promises.readFile(absolutePath, "utf-8")
-
-				// Parse and apply the edit blocks
-				const editBlocks = parseDiffBlocks(diff, absolutePath)
-				newContent = await applyEditBlocksToFile(originalContent, editBlocks)
-			} else {
-				this.logger(`File does not exist or diff is not provided.`, "debug")
-				if (!content) {
-					throw new Error("File does not exist, but 'kodu_content' parameter is missing")
-				}
-				newContent = content
+			if (diff) {
+				return await this.finalizeInlineEdit(relPath, diff)
+			} else if (content) {
+				return await this.finalizeFileEdit(relPath, content)
 			}
-
-			this.logger(`New content: ${newContent}`, "debug")
-			// Show changes in diff view
-			await this.showChangesInDiffView(relPath, newContent)
-			this.logger(`Changes shown in diff view`, "debug")
-
-			this.logger(`Asking for approval to write to file: ${relPath}`, "info")
-			const { response, text, images } = await this.params.ask(
-				"tool",
-				{
-					tool: {
-						tool: "write_to_file",
-						content: newContent,
-						approvalState: "pending",
-						path: relPath,
-						ts: this.ts,
-					},
-				},
-				this.ts
-			)
-
-			if (response !== "yesButtonTapped") {
-				await this.params.updateAsk(
-					"tool",
-					{
-						tool: {
-							tool: "write_to_file",
-							content: newContent,
-							approvalState: "rejected",
-							path: relPath,
-							ts: this.ts,
-							userFeedback: text,
-						},
-					},
-					this.ts
-				)
-				await this.diffViewProvider.revertChanges()
-
-				if (response === "noButtonTapped") {
-					// return formatToolResponse("Write operation cancelled by user.")
-					// return this.toolResponse("rejected", "Write operation cancelled by user.")
-					return this.toolResponse("rejected", "Write operation cancelled by user.")
-				}
-				// If not a yes or no, the user provided feedback (wrote in the input)
-				await this.params.say("user_feedback", text ?? "The user denied this operation.", images)
-				// return formatToolResponse(
-				// 	`The user denied the write operation and provided the following feedback: ${text}`
-				// )
-				return this.toolResponse("feedback", text ?? "The user denied this operation.", images)
-			}
-
-			this.logger(`User approved to write to file: ${relPath}`, "info")
-
-			this.logger(`Saving changes to file: ${relPath}`, "info")
-			// Save changes and handle user edits
-			const { userEdits, finalContent } = await this.diffViewProvider.saveChanges()
-			this.logger(`Changes saved to file: ${relPath}`, "info")
-			this.koduDev.getStateManager().addErrorPath(relPath)
-
-			// Final approval state
-			await this.params.updateAsk(
-				"tool",
-				{
-					tool: {
-						tool: "write_to_file",
-						content: newContent,
-						approvalState: "approved",
-						path: relPath,
-						ts: this.ts,
-					},
-				},
-				this.ts
-			)
-			this.logger(`Final approval state set for file: ${relPath}`, "info")
-			if (userEdits) {
-				await this.params.say(
-					"user_feedback_diff",
-					JSON.stringify({
-						tool: fileExists ? "editedExistingFile" : "newFileCreated",
-						path: getReadablePath(getCwd(), relPath),
-						diff: userEdits,
-					} as ClaudeSayTool)
-				)
-				return this.toolResponse(
-					"success",
-					`The user made the following updates to your content:\n\nThe updated content has been successfully saved to ${relPath.toPosix()}
-					Here is the latest file content:
-					\`\`\`
-					${finalContent}
-					\`\`\`
-					`
-				)
-			}
-
-			let toolMsg = `The content was successfully saved to ${relPath.toPosix()}.
-			Here is the latest file content:
-			\`\`\`
-			${finalContent}
-			\`\`\`
-			`
-			if (detectCodeOmission(this.diffViewProvider.originalContent || "", finalContent)) {
-				this.logger(`Truncated content detected in ${relPath} at ${this.ts}`, "warn")
-				toolMsg = `The content was successfully saved to ${relPath.toPosix()},
-				but it appears that some code may have been omitted. In caee you didn't write the entire content and included some placeholders or omitted critical parts, please try again with the full output of the code without any omissions / truncations anything similar to "remain", "remains", "unchanged", "rest", "previous", "existing", "..." should be avoided.
-				Here is the latest file content:
-				\`\`\`
-				${finalContent}
-				\`\`\``
-			}
-
-			return this.toolResponse("success", toolMsg)
+			throw new Error("Missing required parameter 'kodu_content' or 'kodu_diff'")
 		} catch (error) {
 			this.logger(`Error in processFileWrite: ${error}`, "error")
 			this.params.updateAsk(
@@ -310,6 +522,20 @@ export class FileEditorTool extends BaseAgentTool {
 		} finally {
 			this.isProcessingFinalContent = false
 			this.diffViewProvider.isEditing = false
+		}
+	}
+
+	public override async abortToolExecution(): Promise<void> {
+		if (this.isAbortingTool) {
+			return
+		}
+		this.isAbortingTool = true
+
+		if (this.params.input.kodu_diff) {
+			await this.inlineEditor.rejectChanges()
+			await this.inlineEditor.dispose()
+		} else {
+			await this.diffViewProvider.revertChanges()
 		}
 	}
 
